% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsd.R
\name{popsd}
\alias{popsd}
\title{popsd}
\usage{
popsd(x)
}
\arguments{
\item{x}{Vector of numbers to calculate the population standard deviation of.}
}
\description{
This function allows you to calculate the population standard deviation (as opposed to the sample standard deviation)
Inspired by multicon.
}
\examples{
x = c(4,7,8,9,10)
popsd(x)
}
\keyword{x}
